backend/requirements.txt
asgiref==3.7.2
black==24.2.0
click==8.1.7
dj-database-url==2.1.0
Django==5.0.3
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
